+++ src/scal/rest_server/api_v1/event_handlers_monthly.go
@@ -487,21 +487,21 @@ func UpdateMonthly(req Request) (*Response, error) {
 	) {
 		eventMeta.FieldsMtime["endJd"] = now
 	}
-	// PARAM="day", PARAM_TYPE="int", PARAM_INT=true
+	// PARAM="day", PARAM_TYPE="uint8", PARAM_INT=true
 	if !reflect.DeepEqual(
 		eventModel.Day,
 		lastEventModel.Day,
 	) {
 		eventMeta.FieldsMtime["day"] = now
 	}
-	// PARAM="dayStartSeconds", PARAM_TYPE="int", PARAM_INT=true
+	// PARAM="dayStartSeconds", PARAM_TYPE="uint32", PARAM_INT=true
 	if !reflect.DeepEqual(
 		eventModel.DayStartSeconds,
 		lastEventModel.DayStartSeconds,
 	) {
 		eventMeta.FieldsMtime["dayStartSeconds"] = now
 	}
-	// PARAM="dayEndSeconds", PARAM_TYPE="int", PARAM_INT=true
+	// PARAM="dayEndSeconds", PARAM_TYPE="uint32", PARAM_INT=true
 	if !reflect.DeepEqual(
 		eventModel.DayEndSeconds,
 		lastEventModel.DayEndSeconds,
@@ -712,7 +712,7 @@ func PatchMonthly(req Request) (*Response, error) {
 	{
 		rawValue, ok := patchMap["day"]
 		if ok {
-			// json Unmarshal converts int to float64
+			// json Unmarshal converts uint8 to float64
 			value, typeOk := rawValue.(float64)
 			if !typeOk {
 				return nil, NewError(
@@ -721,7 +721,7 @@ func PatchMonthly(req Request) (*Response, error) {
 					nil,
 				)
 			}
-			eventModel.Day = int(value)
+			eventModel.Day = uint8(value)
 			delete(patchMap, "day")
 			eventMeta.FieldsMtime["day"] = now
 		}
@@ -729,7 +729,7 @@ func PatchMonthly(req Request) (*Response, error) {
 	{
 		rawValue, ok := patchMap["dayStartSeconds"]
 		if ok {
-			// json Unmarshal converts int to float64
+			// json Unmarshal converts uint32 to float64
 			value, typeOk := rawValue.(float64)
 			if !typeOk {
 				return nil, NewError(
@@ -738,7 +738,7 @@ func PatchMonthly(req Request) (*Response, error) {
 					nil,
 				)
 			}
-			eventModel.DayStartSeconds = int(value)
+			eventModel.DayStartSeconds = uint32(value)
 			delete(patchMap, "dayStartSeconds")
 			eventMeta.FieldsMtime["dayStartSeconds"] = now
 		}
@@ -746,7 +746,7 @@ func PatchMonthly(req Request) (*Response, error) {
 	{
 		rawValue, ok := patchMap["dayEndSeconds"]
 		if ok {
-			// json Unmarshal converts int to float64
+			// json Unmarshal converts uint32 to float64
 			value, typeOk := rawValue.(float64)
 			if !typeOk {
 				return nil, NewError(
@@ -755,7 +755,7 @@ func PatchMonthly(req Request) (*Response, error) {
 					nil,
 				)
 			}
-			eventModel.DayEndSeconds = int(value)
+			eventModel.DayEndSeconds = uint32(value)
 			delete(patchMap, "dayEndSeconds")
 			eventMeta.FieldsMtime["dayEndSeconds"] = now
 		}
@@ -1092,7 +1092,7 @@ func MergeMonthly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="day", PARAM_TYPE="int", PARAM_INT=true
+		// PARAM="day", PARAM_TYPE="uint8", PARAM_INT=true
 		inputValue := inputEventModel.Day
 		lastValue := lastEventModel.Day
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -1115,7 +1115,7 @@ func MergeMonthly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="dayStartSeconds", PARAM_TYPE="int", PARAM_INT=true
+		// PARAM="dayStartSeconds", PARAM_TYPE="uint32", PARAM_INT=true
 		inputValue := inputEventModel.DayStartSeconds
 		lastValue := lastEventModel.DayStartSeconds
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -1138,7 +1138,7 @@ func MergeMonthly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="dayEndSeconds", PARAM_TYPE="int", PARAM_INT=true
+		// PARAM="dayEndSeconds", PARAM_TYPE="uint32", PARAM_INT=true
 		inputValue := inputEventModel.DayEndSeconds
 		lastValue := lastEventModel.DayEndSeconds
 		if reflect.DeepEqual(inputValue, lastValue) {
